+++ src/routes/api/v1/user/auth/+server.ts
@@ -1,3 +1,4 @@
+import { json } from '@sveltejs/kit';
 import type { APIUser, RESTPostOAuth2AccessTokenResult } from 'discord-api-types/v10';
 import type { IPartialGuild, TSessionID } from 'src/interfaces';
 import { setSession } from '$lib/utils/sessionHandler';
@@ -7,9 +8,11 @@ import { DISCORD_OAUTH_CLIENT_SECRET } from '$env/static/private';
 import { PUBLIC_DISCORD_OAUTH_CLIENT_ID } from '$env/static/public';
 import { env } from '$env/dynamic/public';
 
-export async function GET({ url }) {
+export async function GET({ url, setHeaders }) {
     const code = url.searchParams.get('code');
-    if (!code) return { status: 400, body: { error: 'No code provided' } };
+    if (!code) return json({ error: 'No code provided' }, {
+        status: 400
+    });
 
     const FormData = new URLSearchParams({
         client_id: PUBLIC_DISCORD_OAUTH_CLIENT_ID,
@@ -37,6 +40,8 @@ export async function GET({ url }) {
         });
 
         const UserData: APIUser = UserRes.data;
+        const Grantdata: RESTPostOAuth2AccessTokenResult = AuthRes.data;
+
 
         // Get the guilds the user is in
         const UserGuildRes = await axios.get(`https://discord.com/api/v10/users/@me/guilds`, {
@@ -54,25 +59,19 @@ export async function GET({ url }) {
         // Optionally, you can upsert the user in the DB here
 
         // Redirect the user and set the session cookie
-        return {
-            status: 302,
-            headers: {
-                'Set-Cookie': cookie.serialize('session_id', SessionID as string, {
-                    path: '/',
-                    httpOnly: true,
-                    sameSite: false,
-                    secure: true,
-                    maxAge: AccessData.expires_in
-                }),
-                Location: '/dashboard'
-            }
-        }
+        setHeaders({
+            'Set-Cookie': cookie.serialize('session_id', SessionID as string, {
+                path: '/',
+                httpOnly: true,
+                sameSite: false,
+                secure: true,
+                maxAge: Grantdata.expires_in
+            })});
+        setHeaders({'location': '/dashboard'});
+        return new Response('', {status: 302 });
 
     } catch (error) {
         console.log(error);
-        return {
-            status: 302,
-            Location: '/authorizationError'
-        }
+        return new Response(undefined, { status: 302 })
     };
 }
