+++ src/routes/api/v1/user/validateSession/+server.ts
@@ -1,3 +1,4 @@
+import { json } from '@sveltejs/kit';
 import { fetchSession } from '$lib/utils/sessionHandler';
 import type { APIUser } from 'discord-api-types/v10';
 import cookie from 'cookie';
@@ -9,15 +10,16 @@ interface IBody {
 export async function POST({ request }) {
     let cookies = cookie.parse(request.headers.get('cookie') || '')
     let sessionId = cookies['session_id'];
-    if (!sessionId) return { status: 400, body: { error: 'Property "sessionId" is required.' } };
+    if (!sessionId) return json({ error: 'Property "sessionId" is required.' }, {
+        status: 400
+    });
 
     const session = fetchSession(sessionId);
-    if (!session) return { status: 400, body: { error: 'Invalid session.' } };
+    if (!session) return json({ error: 'Invalid session.' }, {
+        status: 400
+    });
 
     const {...apiUser}:APIUser = session;
 
-    return {
-        status: 200,
-        body: apiUser
-    };
+    return json(apiUser);
 }
