+++ src/routes/dashboard/+page.ts
@@ -1,0 +1,5 @@
+import type { APIUser } from 'discord-api-types/v10';
+
+export async function load({ }) {
+	return { };
+}
